moves isolated AGSPopup+Sort and AGSPopup+Comparison to caller file, removes redundant cody by passing sort operator as function
